backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.1
python-multipart==0.0.9
python-dotenv==1.0.1
google-generativeai==0.5.4
httpx==0.27.0
sqlalchemy==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
supabase